pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sigmarket"
version = "0.1.0"
description = "PyTorch-based signature-method market simulation library"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
authors = [
    {name = "union"}
]
keywords = ["signature-methods", "market-simulation", "generative-models", "pytorch", "finance"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Financial and Insurance Industry",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Office/Business :: Financial",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "torch>=2.0.0",
    "signatory>=1.2.6",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "pandas-datareader>=0.10.0",
    "yfinance>=0.2.0",
    "pyarrow>=12.0.0",
    "scipy>=1.10.0",
    "scikit-learn>=1.2.0",
    "tqdm>=4.65.0",
    "pyyaml>=6.0",
    "joblib>=1.2.0",
    "esig>=0.10.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.3.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.2.0",
    "isort>=5.12.0",
]
logging = [
    "tensorboard>=2.12.0",
    "wandb>=0.15.0",
]
viz = [
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
]
notebooks = [
    "jupyter>=1.0.0",
    "ipykernel>=6.22.0",
    "ipywidgets>=8.0.0",
]

[project.urls]
Homepage = "https://github.com/union/sigmarket"
Documentation = "https://github.com/union/sigmarket"
Repository = "https://github.com/union/sigmarket"
Issues = "https://github.com/union/sigmarket/issues"

[tool.black]
line-length = 100
target-version = ["py38", "py39", "py310", "py311"]
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | market_simulator
)/
'''

[tool.isort]
profile = "black"
line_length = 100
skip_gitignore = true
extend_skip = ["market_simulator"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "--strict-markers --cov=sigmarket --cov-report=term-missing"

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "esig.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "signatory.*"
ignore_missing_imports = true
